+++ persistent-postgresql/Database/Persist/Postgresql.hs
@@ -1193,7 +1193,7 @@ findAlters defs edef col@(Column name isNull sqltype def _gen _defConstraintName
                 )
 
 -- We check if we should alter a foreign key. This is almost an equality check,
--- except we consider 'Nothing' and 'Just Restrict' equivalent.
+-- except we consider 'Nothing' and 'Just NoAction' equivalent.
 equivalentRef :: Maybe ColumnReference -> Maybe ColumnReference -> Bool
 equivalentRef Nothing Nothing = True
 equivalentRef (Just cr1) (Just cr2) =
@@ -1204,8 +1204,8 @@ equivalentRef (Just cr1) (Just cr2) =
   where
     eqCascade :: Maybe CascadeAction -> Maybe CascadeAction -> Bool
     eqCascade Nothing Nothing         = True
-    eqCascade Nothing (Just Restrict) = True
-    eqCascade (Just Restrict) Nothing = True
+    eqCascade Nothing (Just NoAction) = True
+    eqCascade (Just NoAction) Nothing = True
     eqCascade (Just cs1) (Just cs2)   = cs1 == cs2
     eqCascade _ _                     = False
 equivalentRef _ _ = False